Task Inspector tasks link to Task Graph Inspector
Previously, TaskGraphId was shown as a plain string - with this change,
they are now shown as a hyperlink, so you can get straight from a task
to the task graph it came from.

Secondly, I've updated the "Run Locally" generated output to quote words
consistently and appropriately, and removed unnecessary ';' from end of
lines (which were redundant).

Changes tested locally and working.