Bug 1245254 - Improve run locally message for developers
The current output of 'run locally' only works if you use vagrant and
we're hoping to not need to do so by using this changes.

It also helps because it does not immediately start executing the entry
script but allows for the developer to initiate which takes aways the
"execute and abort" problem.

There is like to be more changes related to this but this is a start.