Bug 1245254 - Remove the need to manually create video devices
In order to run tasks which need a video device to be created for docker, we need
to call 'modprobe v4l2loopback' and have 'v4l2loopback-dkms' installed.

These changes adjust the output of 'Run locally' to check if the package
needs to be installed or if the device needs to be created.

The user will need to enter their root password in order for these
to happen, hence, giving them a change to abort if they don't want to.