User input required commands should run at the very beginning of the script
The commands to initialize video devices can require user input.
However, these commands are to be executed after the downloading of the
docker image which can take a very long time. Requiring user input
mid-script is not good since the user might go and work on other things
while hoping the script would run to completion without their
intervention.

Moving the downloading of the docker image later in the script allows
for user input to be tackled immediately, thus, giving the user the
chance to interact with the script before moving on with other tasks.