Turn src/interactive and src/shell into React classes
Since we are using a single page application, having
`src/interactive/index.js` and `src/shell/index.js`
not in a React class will cause both files to be loaded simultaneously.
In order to avoid that, both files needs to be transformed
into a React class so that React Router only loads the
respective component.

Also, a manual download of utf-8-validate and bufferutil
had to be done to make both terminals work.