Add getWrappedInstance method to each HOC
When using higher-order components, `refs` are not passed through all
props to the wrapped component. If you just use refs, you might end up
referencing the outer container when in fact you are trying to access
the wrapped instance.

To fix this, each HOC is augmented with the method `getWrappedInstance`
that returns its respective wrapped component. For more information, see
https://facebook.github.io/react/docs/higher-order-components.html#refs-arent-passed-through.