Remove the overflow hidden styling on the body element
Previously, in interactive.less and shell.less, there was the following
styling: `body { overflow: hidden }`.

Since taskcluster-tools is now a single page app, all css styles are
loaded in the intial load. As a result, all pages became unscrollable.
To solve this, I removed the overflow style. It doesn't seem to have
caused any effect  on both `/interactive` and `/shell`.