Mega routing refactor (#223)
* Modify gitignore

* Migration init

Migrated Task Inspector, Task Group Inspector & Task Creator

* Migrate quickstart

* Migrate purge-caches

* Migrate Index folder. Deep linking is still in work progress

* Migrate Login and Secrets

* Migrate Status

* Migrate Diagnostics

* Migrate Hooks

* Add deep linking to Index browser

* Migrate auth/clients

* Migrate PulseInspector

* Migrate Credentials

* Migrate auth/scopes

* Make auth/clients use encoded path

* Make auth/roles use encoded path

* Make /secrets use an encoded path

* Nit to src/hooks

* Migrate AwsProvisioner

* Migrate one-click-loaner and one-click-loaner/connect

* Redirect task-graph-inspector and push-inspector to task-group-inspector

Also added small nits

* Make redirects persist optional pathnames

* Modify tools-preset to use src/template.ejs

* Remove unused code

* Add initial implementation of HOCs: TaskClusterEnhance, CreateWebListener, CreateWatchState

* Remove mixins from task-inspector. Will use HOCs instead

* Improve usability of Taskcluster Tools [continued] (#214)

* Remove mixins from task-creator

* Remove mixins from purge-caches

* Make HOCs use reloadOnKeys and reloadOnProps properly

Previously not all HOCs had the functionality of `reloadOnProps`
implemented. In addition, `reloadOnKeys` used to check `this.props.match.params` instead of
`this.state`.

This commit includes the correct implementation of `reloadOnProps` and `reloadOnKeys`

* Make task-inspector use updated HOCs

* Remove mixins from Secrets

* Remove mixins from PulseInspector

* Prettify PulseInspector

* Remove mixins from Index Browser

* Remove mixins from src/lib/ui/artifactlist.jsx

* Remove mixins from src/lib/ui components and refined the HOC

* Removed mixins from:
- src/lib/ui/ConfirmActionMenuItem.jsx
- src/lib/ui/PurgeCacheMenuItem.jsx
- src/lib/ui/RetriggerMenuItem.jsx
- src/lib/ui/artifactlist.jsx
- src/lib/ui/confirmaction.jsx
- src/lib/ui/logview.jsx
- src/lib/ui/previoustasks.jsx
- src/lib/ui/taskinfo.jsx

* Added shouldComponentUpdate logic to the higher order components in
src/lib/utils.jsx

* Remove mixins from src/lib/ui/timeinput.jsx

* Move to neutrino v5

* Remove mixins from aws-provisioner and refine onTaskClusterUpdate

Removed mixins from aws-provisoner. In addition, onTaskClusterUpdate
was modified so that a component state does not update on every
`taskcluster-update` dispatched event.

* Refine TaskClusterEnhance

When credentials change, handleCredentialsChanged is called. Previously,
it was making a call to updateWrappedState and this was causing an
infinite loop reloading the component.

* Make entryview update upon index browser selection

entryview was not triggering `componentDidUpdate` when the namespace
changes. This was fixed by sending `{...this.props}.`

* Start the initial load in cachemanager and secreteditor

* Remove mixins from src/hooks and refine TaskClusterEnhance

`shouldComponentUpdate` inside TaskClusterEnhance  would only
return true if the URL changes. This can become a problem if a callback
requires a rerender without modifying the URL (e.g., modifying component
props). To fix this, `shouldComponentUpdate` of the HOC is removed.

* Rename class method to follow used convention across components

* Turn src/interactive and src/shell into React classes

Since we are using a single page application, having
`src/interactive/index.js` and `src/shell/index.js`
not in a React class will cause both files to be loaded simultaneously.
In order to avoid that, both files needs to be transformed
into a React class so that React Router only loads the
respective component.

Also, a manual download of utf-8-validate and bufferutil
had to be done to make both terminals work.

* Add rule to remove terminal warning regarding 'bindings'

* Remove mixins from one-click-loaner and loaner-button

* Remove mixins from Display

* Add routes path to the rest of components

* Remove DOMContentLoaded event listener after callback

* Add getWrappedInstance method to each HOC

When using higher-order components, `refs` are not passed through all
props to the wrapped component. If you just use refs, you might end up
referencing the outer container when in fact you are trying to access
the wrapped instance.

To fix this, each HOC is augmented with the method `getWrappedInstance`
that returns its respective wrapped component. For more information, see
https://facebook.github.io/react/docs/higher-order-components.html#refs-arent-passed-through.

* Remove mixins from runinfo.jsx and taskview.jsx

* Fix refs problem in src/hooks

* Remove the overflow hidden styling on the body element

Previously, in interactive.less and shell.less, there was the following
styling: `body { overflow: hidden }`.

Since taskcluster-tools is now a single page app, all css styles are
loaded in the intial load. As a result, all pages became unscrollable.
To solve this, I removed the overflow style. It doesn't seem to have
caused any effect  on both `/interactive` and `/shell`.

* Remove mixins from src/auth/clients

* Remove mixins from src/auth/roles

* Remove mixins from src/auth/scopes

* Remove mixins from Credentials

* Fix hrefs in taskinfo.jsx

* Load components asynchronously and refine tools-preset

* Remove unused taskcluster-reload events from components

Components not making  use of `reloadOnProps` and `reloadOnKeys` feature
in TaskClusterEnhance do not need  the event listener
`taskcluster-reload`.

* Add name key to taskclusterOpts of 2 menu items

* Remove the old mixins in src/lib/utils.jsx

* Modify ref to point to wrapped component

* Add jQuery and $ as global

* Apply linting

* Transfer some dependencies to dev dependencies

* add yarn.lock

* Add documentation to each HOC

* Always render renderWaitFor before renderSpinner

* Call load in componentWillMount

* Routing overhaul (#221)

* Routing overhaul

* Linting fixes, deps update, minify fix

* Name async chunks

* Adding hash parameters for line highlighting

* Fixing deprecated eslint message

* Upgrading neutrino preset for shorter bundle directories when building

* Change page titles on transition

* Move the <HelmetTitle>s into the render method

* Nit

* Add blank property to Helmet Title

* Add blank state to Home

* Use TextDecoder to decode Uint8Array

* Add heroku static website configuration

* Add security headers

* Add cache-control headers

* Remove s3 uploading

* Convert TaskCluster to Taskcluster

* Handle redirect

* Remove LegacyAPI from render and put in a different file

* Task redirection with empty ID, fullscreen log view

* Embedding google fonts in CSS

* Adding 404 route handler

* Renaming legacy redirect component

* Taskcluster renames, github client reference error

* Fixing pulse messages and hooks refreshing

* nits