Take out obsolete top-level metadata attribute
@imbstack told me it didn't belong and @djmitche verified that nothing bad
happens when you remove it from a task. I am moderately certain that this
commit removes the spurious metadata section while leaving the necessary one
in place, but someone better at web apps should verify that.

The old thing with an extra metadata field:

    version: 0
    metadata:
      ...
    tasks:
      - provisionerId: '{{ taskcluster.docker.provisionerId }}'
        ...
        payload:
            ...
        metadata:
            ...

The new correct version that this is supposed to emit instead:

    version: 0
    tasks:
      - provisionerId: '{{ taskcluster.docker.provisionerId }}'
        ...
        payload:
            ...
        metadata:
            ...