[interactive] Provide easy to go link for resolved task.
When the task is resolved, the UI provides no link to the task. We have
the task ID in the URL, but besides it is unintuitive (a user hardly
will figure out it), you have to copy the ID, open the task-inspector
and paste it to get task info.

We format the "resolved" message to include the link to the task to get
detailed info about it.