Bug 1395358 - switch to UserSession for tracking user session
Use a UserSession instance nstead of "credentials", since we don't
always have credentials and need to be a little smarter about the
current user.

The "Clients" component is updated to take a UserSession and hand the
appropriate information to the client constructors.

The Credentials Manager is updated to load the credentials
asynchronously and to display the username and, for debugging purposes,
user ID.

The client editor is updated to wait for credentials to load before
displaying, so that it can use the clientId to guess at clientIds it
should display/create.