Fetch client prefix before loading the client (#403)
Presently, on component mount, it is loading the client before even
having the time to fetch the client prefix. As a consequence, tools will
think that it's dealing with a new client, hence providing the client editor.
If the client exist, then they should be prompted with the view of
either resetting the access token or creating a new client.