Add default GitHub events
This ensures .taskcluster.yml files generated with the quick start tool
actually cause Taskcluster jobs to be triggered,
instead of seeming to not do anything.

Because the current job discoverability story is not great
(best way to find jobs is via GitHub status checks),
this should help reduce confusion for newcomers to TaskCluster.