Add documentation to bring taskcluster-tools closer to the Github recommended community standards (#546)
Added licensing info (MPL2), contribution suggestions, and a link to the Mozilla Community Participation Guidelines.

Addresses https://bugzil.la/1408073 for this repo.

Tagging Dustin and Brian for review because I'd like to apply this exact same patch to all our repos if it passes muster. Tagging Eli as the module owner here.